minor fix + a lot of features

+++ src/main/java/me/minhcrafters/noteblockplayer/utils/TimeUtils.java
@@ -40,7 +40,7 @@ public class TimeUtils {
                 time += (long) Integer.parseInt(hourString) * 60 * 60 * 1000;
             }
             time += (long) Integer.parseInt(minuteString) * 60 * 1000;
-            time += Double.parseDouble(secondString) * 1000.0;
+            time += (long) (Double.parseDouble(secondString) * 1000.0);
             return time;
         } else {
             throw new IOException("Invalid time pattern");
